extract point processing into new directory
