+++ src/commands/convert.ts
@@ -29,7 +29,7 @@ export async function convert(filepaths: string[], options: OptionValues) {
   const { multiColor } = options;
 
   const timeInt = Number.parseInt(options.time);
-  const longOffset = Number.parseFloat(options.longOffset) || 0;
+  const longOffset = Number.parseFloat(options.longOffset) % 360 || 0;
 
   const colorMap: RgbColorArrayType[] = createColorArray(
     color,
