+++ src/modules/processPointData/createSvgTypes.ts
