+++ src/modules/processPointData/isFeatureValid.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, test } from 'vitest';
 import { shapeTypes } from '@projectTypes/shapeTypes';
-import { isFeatureValid } from '../createSvg/isFeatureValid';
+import { isFeatureValid } from './isFeatureValid';
 
 describe('isFeatureValid', () => {
   const shapeTypeTestArray = shapeTypes.map((shapeType) => ({
