+++ src/modules/processPointData/isFeatureValid.ts
