+++ src/modules/processPointData/parsePoints.ts
