+++ src/modules/processPointData/shortDistanceCrosses360.test.ts
