+++ src/modules/processPointData/shortDistanceCrosses360.ts
